+++ src/test/java/BaseTest.java
@@ -54,11 +54,7 @@ public class BaseTest {
         driver.quit();
     }
 
-    public String generateRandomName() {
-        Faker faker = new Faker(new Locale("en-US"));
-        String newName = faker.name().firstName();
-        return newName;
-    }
+
 
     public String generateRandomPlaylistName() {
         Faker faker = new Faker(new Locale("en-US"));
