+++ src/test/java/PageModel/Basepage.java
@@ -1,5 +1,6 @@
 package PageModel;
 
+import com.github.javafaker.Faker;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -8,6 +9,7 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 import java.time.Duration;
+import java.util.Locale;
 
 public class Basepage {
 
@@ -26,7 +28,11 @@ public class Basepage {
         driver.navigate().refresh();
 
     }
-
+    public String generateRandomName() {
+        Faker faker = new Faker(new Locale("en-US"));
+        String newName = faker.name().firstName();
+        return newName;
+    }
